python code aangepast, maakt nu ook totaal aan indien het niet bestaat

+++ Python website/svdw.py
@@ -1,5 +1,6 @@
 import csv
 from datetime import datetime
+from os.path import exists
 namen = {"Anton Vanhauwere": 0, "Dries Hofman": 0, "Dries Uytterhaegen": 0, "Emiel van Wetter": 0, "Gilles Speltincx": 0, "Isaak Noerens": 0, "Jari Verhecken": 0, "Jens De Temmerman": 0, "Jens Lateur": 0, "Jona Vangansbeke": 0, "Lucas Ardyns": 0, "Maarten Van Snick": 0, "Marcel Machiels": 0,
          "Mathias Schepers": 0, "Matthias Callebaut": 0, "Matthieu De Pauw": 0, "Mauritz Carlier": 0, "Miel Herreman": 0, "Nisse Verstuyft": 0, "Reggie Demoor": 0, "Sem Backaert": 0, "Thomas Schepers": 0, "Tim Notebaert": 0, "Wout Van der Meulen": 0}
 lineCount = 0
@@ -35,15 +36,22 @@ with open('Speler Van De Week.csv', 'r') as infile:
         writer.writerow('')
         writer.writerow(['Aantal Personen die stemden', lineCount])
 
-with open('SVDWTotaal.csv') as infile:
-    reader = csv.reader(infile)
-    oudTotaal = {row[0]: row[1] for idx,
-                 row in enumerate(reader) if idx in range(0, 24)}
-
-with open('SVDWTotaal.csv', 'w', newline='') as outfile:
-    writer = csv.writer(outfile, delimiter=',')
-    for i, j in oudTotaal.items():
-        oudTotaal[i] = float(j) + float(namen[i])
-    oudTotaal = dict(sorted(oudTotaal.items(), key=lambda item: item[1]))
-    for i, j in oudTotaal.items():
-        writer.writerow([i, ('%.2f' % j)])
+if exists('SVDWTotaal.csv'):
+    with open('SVDWTotaal.csv') as infile:
+        reader = csv.reader(infile)
+        oudTotaal = {row[0]: row[1] for idx,
+                     row in enumerate(reader) if idx in range(0, 24)}
+    with open('SVDWTotaal.csv', 'w', newline='') as outfile:
+        writer = csv.writer(outfile, delimiter=',')
+        for i, j in oudTotaal.items():
+            oudTotaal[i] = float(j) + float(namen[i])
+        oudTotaal = dict(sorted(oudTotaal.items(), key=lambda item: item[1]))
+        for i, j in oudTotaal.items():
+            writer.writerow([i, ('%.2f' % j)])
+else:
+    with open('SVDWTotaal.csv', 'w', newline='') as outfile:
+        writer = csv.writer(outfile, delimiter=',')
+        beginTotaal = {"Anton Vanhauwere": 0, "Dries Hofman": 0, "Dries Uytterhaegen": 0, "Emiel van Wetter": 0, "Gilles Speltincx": 0, "Isaak Noerens": 0, "Jari Verhecken": 0, "Jens De Temmerman": 0, "Jens Lateur": 0, "Jona Vangansbeke": 0, "Lucas Ardyns": 0, "Maarten Van Snick": 0, "Marcel Machiels": 0,
+                       "Mathias Schepers": 0, "Matthias Callebaut": 0, "Matthieu De Pauw": 0, "Mauritz Carlier": 0, "Miel Herreman": 0, "Nisse Verstuyft": 0, "Reggie Demoor": 0, "Sem Backaert": 0, "Thomas Schepers": 0, "Tim Notebaert": 0, "Wout Van der Meulen": 0}
+        for i, j in beginTotaal.items():
+            writer.writerow([i, j])
